qcow_grabber/requirements.txt
beautifulsoup4
python-glanceclient==0.13.1
python-keystoneclient
python-novaclient
requests
